Create category three and update other category else and some...

+++ workbench/king/backend/src/controllers/ContactController.php
@@ -1,6 +1,7 @@
 <?php namespace King\Backend;
 
-use \View;
+use \View,
+    \Request;
 
 class ContactController extends \BaseController{
 
@@ -11,15 +12,26 @@ class ContactController extends \BaseController{
      */
     protected $layout = 'backend::layouts._master';
 
-
+    /**
+     * Display a listing of the resource.
+     *
+     * @return Response
+     */
     public function index(){
 
         $this->layout->content = View::make('backend::contact.index', array(
-            'contacts' => Contact::all(),
+            'contacts' => Contact::orderBy('id', 'desc')->paginate('15'),
             'total' => Contact::count()
         ));
     }
 
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     *
+     * @return Response
+     */
     public function show($id){
 
         $contact = Contact::find($id);
@@ -32,19 +44,50 @@ class ContactController extends \BaseController{
         ));
     }
 
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     *
+     * @return Response
+     */
     public function destroy($id){
 
-        $contact = Contact::find($id);
-        if(is_null($contact)){
-            return _Common::redirectWithMsg('adminErrors', 'Resource does not exist.', '/admin/contacts');
+        if(Request::isMethod('DELETE')){
+
+            $contact = Contact::find($id);
+            if(is_null($contact)){
+                return _Common::redirectWithMsg('adminErrors', 'Resource does not exist.', '/admin/contacts');
+            }
+
+            try{
+                $contact->delete();
+            } catch (Exception $ex) {
+                return _Common::redirectWithMsg('adminErrors', 'Opp! please try again.', '/admin/contacts');
+            }
+
+            return _Common::redirectWithMsg('adminWarning', 'Delete success.', '/admin/contacts');
+        }
+    }
+
+    /**
+     * Remove all resources
+     *
+     * @return response
+     */
+    public function destroyAll(){
+
+        if(Request::isMethod('DELETE')){
+
+            try {
+                Contact::truncate();
+            } catch (Exception $ex) {
+                return _Common::redirectWithMsg('adminErrors', 'Opp! Please try again.', '/admin/contacts');
+            }
+
+            return _Common::redirectWithMsg('adminWarning', 'Delete success.', '/admin/contacts');
         }
 
-        try{
-            $contact->delete();
-        } catch (Exception $ex) {
-            return _Common::redirectWithMsg('adminErrors', 'Opp! please try again.', '/admin/contacts');
-        }
-
-        return _Common::redirectWithMsg('adminWarning', 'Delete success.', '/admin/contacts');
+        $this->layout->content = View::make('backend::contact.delete-all-comfirmation', array());
     }
 }
